Set up debug logging, split up phone controller
