+++ controllers/phone/index.js
@@ -1,0 +1,4 @@
+var newCall = require('./new-call');
+var redirectCall = require('./redirect-call');
+
+module.exports = Object.assign({}, newCall, redirectCall);
